refactor: ChatNotification 도메인 위치 변경(chat -> notification)
