+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/dto/notification/NotificationRespDto.java
@@ -2,7 +2,7 @@ package kr.ac.kumoh.illdang100.tovalley.dto.notification;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
-import kr.ac.kumoh.illdang100.tovalley.domain.chat.ChatNotification;
+import kr.ac.kumoh.illdang100.tovalley.domain.notification.ChatNotification;
 import lombok.AllArgsConstructor;
 import lombok.Data;
