+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/service/notification/NotificationServiceImpl.java
@@ -3,8 +3,8 @@ package kr.ac.kumoh.illdang100.tovalley.service.notification;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kr.ac.kumoh.illdang100.tovalley.domain.chat.ChatNotification;
-import kr.ac.kumoh.illdang100.tovalley.domain.chat.ChatNotificationRepository;
+import kr.ac.kumoh.illdang100.tovalley.domain.notification.ChatNotification;
+import kr.ac.kumoh.illdang100.tovalley.domain.notification.ChatNotificationRepository;
 import kr.ac.kumoh.illdang100.tovalley.domain.chat.ChatRoom;
 import kr.ac.kumoh.illdang100.tovalley.domain.chat.ChatRoomRepository;
 import kr.ac.kumoh.illdang100.tovalley.domain.chat.kafka.Message;
